[CHANGE] #69 - rm old files, less fucntions, more conscise, better efficiency
