+++ headers/input.h
@@ -6,7 +6,7 @@
 /*   By: iatilla- <iatilla-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/26 14:49:06 by schiper           #+#    #+#             */
-/*   Updated: 2025/05/01 18:30:30 by iatilla-         ###   ########.fr       */
+/*   Updated: 2025/05/02 02:44:05 by iatilla-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,5 +43,7 @@ int							read_loop(char **env);
  * @return: Always returns 0
  */
 int							display_tokens(t_token *tokens);
+void						parse_segment_characters(t_parse_params *params,
+								t_parse_state *segment_state);
 
 #endif // INPUT_H
