+++ sources/enviromental/env_dollar_question.c
