+++ sources/input/cleanup.c
